backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic-settings==2.9.1
python-dotenv==1.1.0
starlette==0.46.2
